requirements.txt
opencv-python
ultralytics
pillow
tkcalendar
ttkthemes
